Add node instrumentation
